created MinhasInterasComponent
